提交update命令

+++ internal/commands/develop_environment_util_update.go
@@ -55,6 +55,10 @@ func NewDevelopEnvironmentUtilUpdateCommand() *cobra.Command {
 					}
 				}
 
+				if err := os.RemoveAll(filepath.Join(system.GetRootPath(), "index", extension)); nil != err {
+					return echo.Error("清理就的[%s]索引库失败: %s", extension, err)
+				}
+
 				if err := utils.GzipDecompressWithProgress(cmd.Context(), filepath.Join(system.GetRootPath(), "index"), filename, file, size); nil != err {
 					_ = file.Close()
 					_ = os.Remove(file.Name())
